+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - TBD
 
+## v0.2.0
+### Added
+- Tests for calculations
+### Changed
+- Moved driver functions into struct for a cleaner API and namespacing
+- Improved tests
+
 ## v0.1.0
 ### Added
 - Project setup, including: